YEAH

+++ bonus/sphere.cpp
@@ -24,8 +24,8 @@ bool Sphere::hit(const Camera ray, float &old_dist)
     float		delta = B * B - dist * dist + this->radius * this->radius;
     if (delta < 0)
         return (0);
-    float       t0 = B - sqrtf(delta);
-    float       t1 = B + sqrtf(delta);
+    float       t0 = B - sqrt(delta);
+    float       t1 = B + sqrt(delta);
     if (t0 < t1 && (t0 > 0.0) && (t0 < old_dist))
     {
         old_dist = t0;
